refactor: add button ref
