+++ packages/client/src/components/common/Button/Button.tsx
@@ -1,5 +1,5 @@
 import { NormalColorType } from '~/styles/';
-import { ButtonHTMLAttributes } from 'react';
+import { ButtonHTMLAttributes, forwardRef } from 'react';
 import * as S from './Button.styles';
 import Ripple from './Ripple';
 
@@ -10,19 +10,15 @@ export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   shadow?: boolean;
 }
 
-const Button = ({
-  size = 'md',
-  color = 'primary',
-  children,
-  shadow = false,
-  ...options
-}: ButtonProps) => {
+const Button = forwardRef<HTMLButtonElement, ButtonProps>(function Button(
+  { size = 'md', color = 'primary', children, shadow = false, ...options },
+  ref,
+) {
   return (
-    <S.Container size={size} color={color} shadow={shadow} {...options}>
+    <S.Container size={size} color={color} shadow={shadow} {...options} ref={ref}>
       <Ripple />
       {children}
     </S.Container>
   );
-};
-
+});
 export default Button;
