+++ packages/client/src/hooks/queries/room/useCreateRoom.ts
@@ -1,6 +1,6 @@
 import RoomAPI from '~/api/room';
 import { useMutation } from '@tanstack/react-query';
-import type { UseMutationOptionsOf } from '~/hooks/types';
+import type { UseMutationOptionsOf } from '~/hooks/queries/types';
 
 const useCreateRoom = (options: UseMutationOptionsOf<typeof RoomAPI.createRoom> = {}) => {
   return useMutation(RoomAPI.createRoom, options);
