+++ packages/client/src/hooks/queries/room/useDeleteRoom.ts
@@ -1,6 +1,6 @@
 import RoomAPI from '~/api/room';
 import { useMutation } from '@tanstack/react-query';
-import type { UseMutationOptionsOf } from '~/hooks/types';
+import type { UseMutationOptionsOf } from '~/hooks/queries/types';
 
 const useDeleteRoom = (options: UseMutationOptionsOf<typeof RoomAPI.deleteRoom> = {}) => {
   return useMutation(RoomAPI.deleteRoom, options);
