+++ packages/client/src/hooks/queries/room/useGetRoom.ts
@@ -1,6 +1,6 @@
 import RoomAPI from '~/api/room';
 import { useQuery } from '@tanstack/react-query';
-import type { UseQueryOptionsOf } from '~/hooks/types';
+import type { UseQueryOptionsOf } from '~/hooks/queries/types';
 
 const useGetRoom = (roomId: string, options: UseQueryOptionsOf<typeof RoomAPI.getRoom> = {}) => {
   return useQuery(getKey(roomId), fetcher(roomId), options);
