+++ packages/client/src/hooks/queries/room/useGetRooms.ts
@@ -1,6 +1,6 @@
 import RoomAPI from '~/api/room';
 import { useQuery } from '@tanstack/react-query';
-import type { UseQueryOptionsOf } from '~/hooks/types';
+import type { UseQueryOptionsOf } from '~/hooks/queries/types';
 
 const useGetRooms = (options: UseQueryOptionsOf<typeof RoomAPI.getRooms> = {}) => {
   return useQuery(getKey(), fetcher(), options);
